사용자가 자신의 정보만 수정 삭제 가능하도록 제한

+++ app/src/main/java/com/codesoom/assignment/controllers/UserController.java
@@ -1,11 +1,14 @@
 package com.codesoom.assignment.controllers;
 
+import com.codesoom.assignment.application.AuthenticationService;
 import com.codesoom.assignment.application.UserService;
 import com.codesoom.assignment.domain.User;
 import com.codesoom.assignment.dto.UserModificationData;
 import com.codesoom.assignment.dto.UserRegistrationData;
 import com.codesoom.assignment.dto.UserResultData;
+import com.codesoom.assignment.errors.UserForbiddenException;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -16,8 +19,11 @@ import javax.validation.Valid;
 public class UserController {
     private final UserService userService;
 
-    public UserController(UserService userService) {
+    private final AuthenticationService authenticationService;
+
+    public UserController(UserService userService, AuthenticationService authenticationService) {
         this.userService = userService;
+        this.authenticationService = authenticationService;
     }
 
     @PostMapping
@@ -28,18 +34,31 @@ public class UserController {
     }
 
     @PatchMapping("{id}")
+    @PreAuthorize("isAuthenticated()")
     UserResultData update(
+            @RequestAttribute Long userId,
             @PathVariable Long id,
             @RequestBody @Valid UserModificationData modificationData
     ) {
-        User user = userService.updateUser(id, modificationData);
-        return getUserResultData(user);
+        if (userId.equals(id)) {
+            User user = userService.updateUser(id, modificationData);
+            return getUserResultData(user);
+        } else {
+            throw new UserForbiddenException();
+        }
     }
 
     @DeleteMapping("{id}")
+    @PreAuthorize("isAuthenticated()")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    void destroy(@PathVariable Long id) {
-        userService.deleteUser(id);
+    void destroy(
+            @RequestAttribute Long userId,
+            @PathVariable Long id
+    ) {
+        if (!userId.equals(id)) {
+            userService.deleteUser(id);
+            throw new UserForbiddenException();
+        }
     }
 
     private UserResultData getUserResultData(User user) {
